requirements.txt
aiofiles==23.2.1
asgiref==3.8.1
asttokens==2.4.1
certifi==2024.2.2
charset-normalizer==3.3.2
decorator==5.1.1
django==5.0.4
django-rich==1.8.0
django-stubs==4.2.7
django-stubs-ext==4.2.7
executing==2.0.1
gunicorn==22.0.0
idna==3.6
ipython==8.23.0
jedi==0.19.1
markdown-it-py==3.0.0
matplotlib-inline==0.1.6
mdurl==0.1.2
packaging==24.0
parso==0.8.4
pexpect==4.9.0
pillow==10.2.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pygments==2.17.2
requests==2.31.0
rich==13.7.1
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.3
traitlets==5.14.2
types-pytz==2024.1.0.20240203
types-pyyaml==6.0.12.20240311
typing-extensions==4.11.0
urllib3==2.2.1
uuid==1.30
wcwidth==0.2.13
